+++ benchmarks/aa_overhead_ocs.py
@@ -3,7 +3,7 @@ from time import perf_counter
 from typing import Any, Union
 
 import matplotlib.pyplot as plt
-from ae_overhead_common import measure_overhead, os_id
+from aa_overhead_common import measure_overhead, os_id
 
 from nc_py_api import Nextcloud, NextcloudApp
